:recycle: move to infra/repo/db.go
